changed coordinates to output radius of curvature and import change

+++ gnss/util/ring_buffer.py
@@ -15,14 +15,14 @@ class RingBuffer(object):
     
     def get(self, size=None):
         if size:
-            indices = (self.index - size + arange(size)) % self.size
+            indices = ((self.index - size + arange(size)) % self.size).astype(int)
             return self.buffer[indices]
-        indices = (self.index + arange(self.size)) % self.size
+        indices = ((self.index + arange(self.size)) % self.size).astype(int)
         return self.buffer[indices]
     
     def extend(self, vals):
         vals = asarray([vals])
-        indices = (self.index + arange(len(vals))) % self.size
+        indices = ((self.index + arange(len(vals))) % self.size).astype(int)
         self.buffer[indices] = vals
         self.index = indices[-1] + 1
